Research poster
A poster which aims to encapsulate some of the formative ideas I have been having in relation to my intended thesis. The statements outlined in the fug of Jean Baudrillard's cigarette smoke are stream of consciousness thoughts I had whilst beginning this task in Photoshop, in relation to the texts I have been reading.