Contextual_Review_Hyperreal_scene.md
# Photorealistic render found on reddit 

## created by u/Hary1495, * https://www.reddit.com/r/toptalent/comments/kyiavv/i_work_with_cgi_and_this_is_my_2nd_attempt_in/* 
 

At first glance, there is something entirely plausible about this scene; its profusion of objects, natural lighting, gratifyingly realistic materials... even its haphazard throws and billowing curtains imply a lived-in-ness, the fleeting impressions of mliving bodies and their occupation of and interaction with a [hyperreal] space. Had I not seen the r/vfx moniker listed against the image’s title as I scrolled past it, I likely would have mistaken this image as a photograph. And yet – when you look a little closer, there is so much amiss, so much that confuses, so many aspects which are nonsensical. 

One such item is the foreground throw, which is prettily splayed against (?), over (?) the corner of the couch, impossibly gripping its side and the nearby cushion. It is a wool blend – its knotted lines of fabric are distinct – a denser fabric than most, and much more so than that of the diaphanous blinds, yet it clings vertically to the edge of the couch, cut through with sweeping folds that would surely collapse it under its own weight and gravity. 

Look closer still, and there are multiple couches in this scene. Multiple couches?! As well adjacent armchairs. On noticing this, I looked for the telltale outline of a grandiose background mirror – its gilt frame, a reflection or curious glint of light playing on its surface – or maybe (the mind gropes) it is a full-length wall mirror – but then where are its joins? There isn’t a mirror. The scene boasts multiple couches, displayed back-to-back, each with its own distinct material effects, casually-impossibly draped throw, gently sculpted pillows – each of which show little sign of indentation or use beyond that of a cloth sim. The distant (blue) armchair is also much more simply rendered. The faux-suede chair closer to the foreground has sagging fabric that outlines a depression made in it by its sitters, though of course, the far chair could be new, less often used, etc – the blue chair’s giveaway is its far too prominent normal detail: the false impression of height and/or depth created by a normal map. 

Off to the right is a wardrobe, mid-century in style. Why? Is this a one-bedroom apartment? An open plan living space with a wardrobe for coats or similar near an unseen front door? If it’s the former, why all the couches? Even the most bourgeois of one-beds would likely not require so much seating (who would it be for?). If it’s the latter, why not something more discreet and stylistically apropos, like a hall stand? 

I’m not even going to analyse the barista-perfect latte and its cutified heart of steam, nor the smattering of ‘children’s toys’, which up close are bizarre, utility-less things, without real form or clarity of purpose. Memeified objects. 

The lighting and general material work in this piece are outstanding though. It bears many of the hallmarks of a photographic representation, though where this a photograph, it truly would have effaced itself of its subject in some uncanny way. 

I’m reminded of that quote by Baudrillard, 

>Similitude is a dream and must remain one, in order for a modicum of illusion and a stage of  the imaginary to exist. One must never pass over to the side of the real, the side of the exact resemblance of the world to itself, of the subject to itself. Because then the image disappears.<